+++ trace/topic_gtrace.go
@@ -406,9 +406,9 @@ func (t *Topic) Compose(x *Topic, opts ...TopicComposeOption) *Topic {
 		}
 	}
 	{
-		h1 := t.OnReaderPublicPopBatchTx
-		h2 := x.OnReaderPublicPopBatchTx
-		ret.OnReaderPublicPopBatchTx = func(t TopicReaderPopBatchTxStartInfo) func(TopicReaderPopBatchTxDoneInfo) {
+		h1 := t.OnReaderPopBatchTx
+		h2 := x.OnReaderPopBatchTx
+		ret.OnReaderPopBatchTx = func(t TopicReaderPopBatchTxStartInfo) func(TopicReaderPopBatchTxDoneInfo) {
 			if options.panicCallback != nil {
 				defer func() {
 					if e := recover(); e != nil {
@@ -424,6 +424,146 @@ func (t *Topic) Compose(x *Topic, opts ...TopicComposeOption) *Topic {
 				r1 = h2(t)
 			}
 			return func(t TopicReaderPopBatchTxDoneInfo) {
+				if options.panicCallback != nil {
+					defer func() {
+						if e := recover(); e != nil {
+							options.panicCallback(e)
+						}
+					}()
+				}
+				if r != nil {
+					r(t)
+				}
+				if r1 != nil {
+					r1(t)
+				}
+			}
+		}
+	}
+	{
+		h1 := t.OnReaderStreamPopBatchTx
+		h2 := x.OnReaderStreamPopBatchTx
+		ret.OnReaderStreamPopBatchTx = func(t TopicReaderStreamPopBatchTxStartInfo) func(TopicReaderStreamPopBatchTxDoneInfo) {
+			if options.panicCallback != nil {
+				defer func() {
+					if e := recover(); e != nil {
+						options.panicCallback(e)
+					}
+				}()
+			}
+			var r, r1 func(TopicReaderStreamPopBatchTxDoneInfo)
+			if h1 != nil {
+				r = h1(t)
+			}
+			if h2 != nil {
+				r1 = h2(t)
+			}
+			return func(t TopicReaderStreamPopBatchTxDoneInfo) {
+				if options.panicCallback != nil {
+					defer func() {
+						if e := recover(); e != nil {
+							options.panicCallback(e)
+						}
+					}()
+				}
+				if r != nil {
+					r(t)
+				}
+				if r1 != nil {
+					r1(t)
+				}
+			}
+		}
+	}
+	{
+		h1 := t.OnReaderUpdateOffsetsInTransaction
+		h2 := x.OnReaderUpdateOffsetsInTransaction
+		ret.OnReaderUpdateOffsetsInTransaction = func(t TopicReaderOnUpdateOffsetsInTransactionStartInfo) func(TopicReaderOnUpdateOffsetsInTransactionDoneInfo) {
+			if options.panicCallback != nil {
+				defer func() {
+					if e := recover(); e != nil {
+						options.panicCallback(e)
+					}
+				}()
+			}
+			var r, r1 func(TopicReaderOnUpdateOffsetsInTransactionDoneInfo)
+			if h1 != nil {
+				r = h1(t)
+			}
+			if h2 != nil {
+				r1 = h2(t)
+			}
+			return func(t TopicReaderOnUpdateOffsetsInTransactionDoneInfo) {
+				if options.panicCallback != nil {
+					defer func() {
+						if e := recover(); e != nil {
+							options.panicCallback(e)
+						}
+					}()
+				}
+				if r != nil {
+					r(t)
+				}
+				if r1 != nil {
+					r1(t)
+				}
+			}
+		}
+	}
+	{
+		h1 := t.OnReaderTransactionCompleted
+		h2 := x.OnReaderTransactionCompleted
+		ret.OnReaderTransactionCompleted = func(t TopicReaderTransactionCompletedStartInfo) func(TopicReaderTransactionCompletedDoneInfo) {
+			if options.panicCallback != nil {
+				defer func() {
+					if e := recover(); e != nil {
+						options.panicCallback(e)
+					}
+				}()
+			}
+			var r, r1 func(TopicReaderTransactionCompletedDoneInfo)
+			if h1 != nil {
+				r = h1(t)
+			}
+			if h2 != nil {
+				r1 = h2(t)
+			}
+			return func(t TopicReaderTransactionCompletedDoneInfo) {
+				if options.panicCallback != nil {
+					defer func() {
+						if e := recover(); e != nil {
+							options.panicCallback(e)
+						}
+					}()
+				}
+				if r != nil {
+					r(t)
+				}
+				if r1 != nil {
+					r1(t)
+				}
+			}
+		}
+	}
+	{
+		h1 := t.OnReaderTransactionRollback
+		h2 := x.OnReaderTransactionRollback
+		ret.OnReaderTransactionRollback = func(t TopicReaderTransactionRollbackStartInfo) func(TopicReaderTransactionRollbackDoneInfo) {
+			if options.panicCallback != nil {
+				defer func() {
+					if e := recover(); e != nil {
+						options.panicCallback(e)
+					}
+				}()
+			}
+			var r, r1 func(TopicReaderTransactionRollbackDoneInfo)
+			if h1 != nil {
+				r = h1(t)
+			}
+			if h2 != nil {
+				r1 = h2(t)
+			}
+			return func(t TopicReaderTransactionRollbackDoneInfo) {
 				if options.panicCallback != nil {
 					defer func() {
 						if e := recover(); e != nil {
@@ -904,8 +1044,8 @@ func (t *Topic) onReaderUpdateToken(o OnReadUpdateTokenStartInfo) func(OnReadUpd
 		return res
 	}
 }
-func (t *Topic) onReaderPublicPopBatchTx(t1 TopicReaderPopBatchTxStartInfo) func(TopicReaderPopBatchTxDoneInfo) {
-	fn := t.OnReaderPublicPopBatchTx
+func (t *Topic) onReaderPopBatchTx(t1 TopicReaderPopBatchTxStartInfo) func(TopicReaderPopBatchTxDoneInfo) {
+	fn := t.OnReaderPopBatchTx
 	if fn == nil {
 		return func(TopicReaderPopBatchTxDoneInfo) {
 			return
@@ -914,6 +1054,66 @@ func (t *Topic) onReaderPublicPopBatchTx(t1 TopicReaderPopBatchTxStartInfo) func
 	res := fn(t1)
 	if res == nil {
 		return func(TopicReaderPopBatchTxDoneInfo) {
+			return
+		}
+	}
+	return res
+}
+func (t *Topic) onReaderStreamPopBatchTx(t1 TopicReaderStreamPopBatchTxStartInfo) func(TopicReaderStreamPopBatchTxDoneInfo) {
+	fn := t.OnReaderStreamPopBatchTx
+	if fn == nil {
+		return func(TopicReaderStreamPopBatchTxDoneInfo) {
+			return
+		}
+	}
+	res := fn(t1)
+	if res == nil {
+		return func(TopicReaderStreamPopBatchTxDoneInfo) {
+			return
+		}
+	}
+	return res
+}
+func (t *Topic) onReaderUpdateOffsetsInTransaction(t1 TopicReaderOnUpdateOffsetsInTransactionStartInfo) func(TopicReaderOnUpdateOffsetsInTransactionDoneInfo) {
+	fn := t.OnReaderUpdateOffsetsInTransaction
+	if fn == nil {
+		return func(TopicReaderOnUpdateOffsetsInTransactionDoneInfo) {
+			return
+		}
+	}
+	res := fn(t1)
+	if res == nil {
+		return func(TopicReaderOnUpdateOffsetsInTransactionDoneInfo) {
+			return
+		}
+	}
+	return res
+}
+func (t *Topic) onReaderTransactionCompleted(t1 TopicReaderTransactionCompletedStartInfo) func(TopicReaderTransactionCompletedDoneInfo) {
+	fn := t.OnReaderTransactionCompleted
+	if fn == nil {
+		return func(TopicReaderTransactionCompletedDoneInfo) {
+			return
+		}
+	}
+	res := fn(t1)
+	if res == nil {
+		return func(TopicReaderTransactionCompletedDoneInfo) {
+			return
+		}
+	}
+	return res
+}
+func (t *Topic) onReaderTransactionRollback(t1 TopicReaderTransactionRollbackStartInfo) func(TopicReaderTransactionRollbackDoneInfo) {
+	fn := t.OnReaderTransactionRollback
+	if fn == nil {
+		return func(TopicReaderTransactionRollbackDoneInfo) {
+			return
+		}
+	}
+	res := fn(t1)
+	if res == nil {
+		return func(TopicReaderTransactionRollbackDoneInfo) {
 			return
 		}
 	}
@@ -1205,16 +1405,83 @@ func TopicOnReaderUpdateToken(t *Topic, readerConnectionID string) func(tokenLen
 }
 
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
-func TopicOnReaderPublicPopBatchTx(t *Topic, c *context.Context, readerConnectionID string, transactionSessionID string, transactionID string) func(error) {
+func TopicOnReaderPopBatchTx(t *Topic, c *context.Context, readerID int64, transactionSessionID string, transactionID string) func(startOffset int64, endOffset int64, messagesCount int, _ error) {
 	var p TopicReaderPopBatchTxStartInfo
 	p.Context = c
+	p.ReaderID = readerID
+	p.TransactionSessionID = transactionSessionID
+	p.TransactionID = transactionID
+	res := t.onReaderPopBatchTx(p)
+	return func(startOffset int64, endOffset int64, messagesCount int, e error) {
+		var p TopicReaderPopBatchTxDoneInfo
+		p.StartOffset = startOffset
+		p.EndOffset = endOffset
+		p.MessagesCount = messagesCount
+		p.Error = e
+		res(p)
+	}
+}
+
+// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
+func TopicOnReaderStreamPopBatchTx(t *Topic, c *context.Context, readerID int64, readerConnectionID string, transactionSessionID string, transactionID string) func(error) {
+	var p TopicReaderStreamPopBatchTxStartInfo
+	p.Context = c
+	p.ReaderID = readerID
 	p.ReaderConnectionID = readerConnectionID
 	p.TransactionSessionID = transactionSessionID
 	p.TransactionID = transactionID
-	res := t.onReaderPublicPopBatchTx(p)
+	res := t.onReaderStreamPopBatchTx(p)
 	return func(e error) {
-		var p TopicReaderPopBatchTxDoneInfo
+		var p TopicReaderStreamPopBatchTxDoneInfo
 		p.Error = e
+		res(p)
+	}
+}
+
+// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
+func TopicOnReaderUpdateOffsetsInTransaction(t *Topic, c *context.Context, readerID int64, readerConnectionID string, transactionSessionID string, transactionID string) func(error) {
+	var p TopicReaderOnUpdateOffsetsInTransactionStartInfo
+	p.Context = c
+	p.ReaderID = readerID
+	p.ReaderConnectionID = readerConnectionID
+	p.TransactionSessionID = transactionSessionID
+	p.TransactionID = transactionID
+	res := t.onReaderUpdateOffsetsInTransaction(p)
+	return func(e error) {
+		var p TopicReaderOnUpdateOffsetsInTransactionDoneInfo
+		p.Error = e
+		res(p)
+	}
+}
+
+// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
+func TopicOnReaderTransactionCompleted(t *Topic, c *context.Context, readerID int64, readerConnectionID string, transactionSessionID string, transactionID string, transactionResult error) func() {
+	var p TopicReaderTransactionCompletedStartInfo
+	p.Context = c
+	p.ReaderID = readerID
+	p.ReaderConnectionID = readerConnectionID
+	p.TransactionSessionID = transactionSessionID
+	p.TransactionID = transactionID
+	p.TransactionResult = transactionResult
+	res := t.onReaderTransactionCompleted(p)
+	return func() {
+		var p TopicReaderTransactionCompletedDoneInfo
+		res(p)
+	}
+}
+
+// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
+func TopicOnReaderTransactionRollback(t *Topic, c *context.Context, readerID int64, readerConnectionID string, transactionSessionID string, transactionID string) func(rollbackError error) {
+	var p TopicReaderTransactionRollbackStartInfo
+	p.Context = c
+	p.ReaderID = readerID
+	p.ReaderConnectionID = readerConnectionID
+	p.TransactionSessionID = transactionSessionID
+	p.TransactionID = transactionID
+	res := t.onReaderTransactionRollback(p)
+	return func(rollbackError error) {
+		var p TopicReaderTransactionRollbackDoneInfo
+		p.RollbackError = rollbackError
 		res(p)
 	}
 }